+++ src/main/java/roomescape/presentation/member/api/MemberController.java
@@ -12,7 +12,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.AccessToken;
+import roomescape.config.AccessToken;
 import roomescape.business.service.member.MemberService;
 import roomescape.config.LoginMember;
 import roomescape.presentation.member.dto.LoginCheckResponseDto;
